ui: promote Users + Quotas to top-level admin nav, consolidate config behind /settings sidebar
The Control Panel dropdown had become a junk drawer of nine items mixing
high-frequency admin work (Users, Quotas) with workspace configuration
(Email, Gopher, Network, S3, GPU hosts) and personal stuff (Keys,
Account). Splits into three concerns:

  - Top-level admin nav: Dashboard, Provision, My machines, Users,
    Quotas, plus S3/GPU when active. Surfaces the daily-driver pages
    instead of hiding them behind a click.

  - /settings page with left sidebar: Sign-in & access, Email,
    Gopher Tunnels, Network, S3 Storage, GPU hosts. Same content the
    admin used to reach from individual dropdown items, now grouped
    under one route. Each existing page mounts inside SettingsLayout
    so the sidebar travels with the operator. Old standalone routes
    (/email, /gophers, /network, /s3, /gpu-host) redirect to their
    /settings/* counterpart so bookmarks resolve.

  - Right user dropdown: Settings, Keys, Account, Sign out. Matches
    the member dropdown shape (Account + Sign out) — admin just gets
    Settings + Keys above it. No more "Control Panel" section header.

Also splits the old /users page in half. /users now hosts the user
table only; the OAuth provider configuration, access code, and
passwordless toggle moved to /settings/sign-in. Users.tsx shrank
from ~2100 lines to ~700 by losing helpers that only the providers
panel used (ProviderPanel, GoogleDomainsSection, GitHubOrgsSection,
AccessCodePanel, OAuthProviderModal, RedirectURIHint, EyeIcon,
CopyIcon, RefreshIcon, SetupNotes, PencilIcon, ProvidersSummary).
Those live in pages/settings/SignIn.tsx now since SettingsSignIn
is their only consumer.

Members get no behaviour change — top nav stays Provision / My
machines / Keys + Account dropdown. /admin remains admin-only.
